rename to pbr_explorer (add underscore) for better readability
