+++ src/reports/rpt_pbr_explorer/example/colorado_simplified.geojson
