+++ src/reports/rpt_pbr_explorer/templates/body.html
